Refactor module part parsers to seperate package
